submit_post classified into new controller
