+++ public/index.php
@@ -10,7 +10,6 @@ $router = new Router();
 
 // Add the routes
 $router->add('', ['controller' => 'HomepageController', 'action' => 'index']);
-$router->add('pesto', ['controller' => 'HomepageController', 'action' => 'show']);
 $router->add('auth', ['controller' => 'AuthenticationController', 'action' => 'authenticate']);
 $router->add('profile', ['controller' => 'ProfileController', 'action' => 'index']);
 $router->add('user_closed', ['controller' => 'UserClosedController', 'action' => 'index']);
@@ -19,7 +18,7 @@ $router->add('ajax_load', ['controller' => 'AjaxLoadPostController', 'action' =>
 $router->add('confirm_post', ['controller' => 'ConfirmPostController', 'action' => 'confirmPost']);
 $router->add('comment_frame', ['controller' => 'CommentController', 'action' => 'frameComment']);
 $router->add('post', ['controller' => 'PostsController', 'action' => 'loadPostsFriends']);
-$router->add('submit', ['controller' => 'PostsController', 'action' => 'submitPost']);
+$router->add('submit_post', ['controller' => 'SubmitPostController', 'action' => 'submitPost']);
 $router->add('logged_out', ['controller' => 'LogoutController', 'action' => 'logout']);
 $router->add('{controller}/{action}');
 $router->add('{controller}/{id:\d+}/{action}');
