+++ src/Controller/HomePageController.php
@@ -3,8 +3,8 @@
 namespace App\Controller;
 
 use App\Repository\UsersRepository;
+use App\Controller\SubmitPostController;
 use Core\Template;
-use App\Controller\PostsController;
 
 Class HomepageController
 {
@@ -16,7 +16,7 @@ Class HomepageController
             header("Location: /auth");
         }
         if(isset($_POST['post'])){
-            $post = new PostsController();
+            $post = new SubmitPostController();
             $post->submitPost($_POST['post_text']);
             header("Location: /"); //Stops the form resubmitting on refresh (duplicate announcement prevention).
         }
